feat: Added the InnerSource domain
This commit adds an initial version of the InnerSource domain.
It is based on the SAP-internal InnerSource domain structure.
The internal references were replaced by external references.

There was a section called "Learn about tooling and Source
Control" on the red belt that was removed. It is a relevant
section, but it was removed as I did not have time to curate
public resources for it, so it is out of the initial version.