feat: Development environment with containers
This commit enables the page to be served locally with
containers.

I removed the notion of working with two "windows" (a jekyll
one and another one to perform git operations), as I assume
this is common knowledge.

I also added the usage of live reload, to facilitate the
testing of changes.

I could not understand the comment "Sometimes the window will
stick", so I removed it.

When compared to our InnerSource counterpart, I also removed
the `version` property on `docker-compose.yml` due to a
warning (see https://github.com/docker/compose/issues/11628).

I assume this does not break compatibility with other
container engines.

This was tested with Docker Desktop 4.29.0 (145265) for Mac.