chore: Removed ignored files
This commit removes tracked files that were already part of
`.gitignore`.
Tracking those files are not needed or valuable for either
GitHub pages deployment or local development.
I also added a comment explaining why we do not track
Gemfile.lock, since it seems to be a deviation of standard
Roby practices (assumption).